move field_names tests to field_names
Also clears up some variable names in match statements.
